+++ src/Language/Haskell/GhclibParserEx/GHC/Settings/Config.hs
@@ -63,8 +63,11 @@ fakeSettings = Settings
                     }
 #endif
     platform =
+#if defined (GHCLIB_API_HEAD)
+      genericPlatform
+#else
       Platform{
-#if defined (GHCLIB_API_HEAD) || defined(GHCLIB_API_902) || defined (GHCLIB_API_900)
+#if defined(GHCLIB_API_902) || defined (GHCLIB_API_900)
     -- It doesn't matter what values we write here as these fields are
     -- not referenced for our purposes. However the fields are strict
     -- so we must say something.
@@ -75,15 +78,12 @@ fakeSettings = Settings
       , platformIsCrossCompiling=False
       , platformLeadingUnderscore=False
       , platformTablesNextToCode=False
-#if defined (GHCLIB_API_HEAD)
-      , platform_constants = Nothing
-#endif
-#if !defined(GHCLIB_API_HEAD) && !defined (GHCLIB_API_900)
+#if !defined (GHCLIB_API_900)
       , platformConstants=platformConstants
 #endif
       ,
 #endif
-#if defined (GHCLIB_API_HEAD) || defined(GHCLIB_API_902)
+#if defined(GHCLIB_API_902)
         platformWordSize=PW8
       , platformArchOS=ArchOS {archOS_arch=ArchUnknown, archOS_OS=OSUnknown}
 #elif defined (GHCLIB_API_810) || defined (GHCLIB_API_900)
@@ -95,6 +95,7 @@ fakeSettings = Settings
 #endif
       , platformUnregisterised=True
       }
+#endif
 #if !defined(GHCLIB_API_HEAD)
     platformConstants =
       PlatformConstants{pc_DYNAMIC_BY_DEFAULT=False,pc_WORD_SIZE=8}